# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kapsalon Vilani API",
        "description": "API for managing salon users, services, and appointments",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bookings/{bookingId}/products": {
            "get": {
                "tags": [
                    "Booking Products"
                ],
                "summary": "Get products for a specific booking",
                "description": "Returns a list of products associated with a specific booking",
                "operationId": "getProductsByBooking",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "description": "ID of booking to get products for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "123e4567-e89b-12d3-a456-426614174000"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Hair Shampoo"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/bookings/{bookingId}/services": {
            "get": {
                "tags": [
                    "Booking Services"
                ],
                "summary": "Get services for a specific booking",
                "description": "Returns a list of services associated with a specific booking",
                "operationId": "getServicesByBooking",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "description": "ID of booking to get services for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "123e4567-e89b-12d3-a456-426614174000"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Haircut"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products/{productId}/categories": {
            "get": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Get all categories for a product",
                "description": "Returns a list of categories associated with a specific product",
                "operationId": "getProductCategories",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID of product to get categories for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of product categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "123e4567-e89b-12d3-a456-426614174000"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Hair Care"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}